requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
boto3==1.35.0
botocore==1.35.99
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
eval_type_backport==0.2.2
fastapi==0.115.0
filetype==1.2.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.43
langchain-openai==0.1.23
langchain-postgres==0.0.15
langchain-text-splitters==0.2.4
langdetect==1.0.9
langgraph==0.2.16
langgraph-checkpoint==1.0.12
langgraph-checkpoint-postgres==1.0.0
langsmith==0.1.147
lxml==6.0.2
marshmallow==3.26.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.109.1
orjson==3.11.3
packaging==24.2
pgvector==0.2.5
propcache==0.4.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.12
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.38.6
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
unstructured==0.11.8
unstructured-client==0.28.1
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
